backend/requirements.txt
# backend/requirements.txt
# Backend ZeeXClub - Dépendances Python

# Framework Web
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Client Supabase
supabase>=2.0.0
postgrest>=0.13.0

# Bot Telegram
pyrogram>=2.0.0
tgcrypto>=1.2.0

# HTTP et API
aiohttp>=3.8.0
requests>=2.31.0

# Utilitaires
python-dotenv>=1.0.0
Pillow>=10.0.0

# Serveur WSGI/ASGI
gunicorn>=21.0.0
uvicorn>=0.23.0

# Cache et performance
redis>=5.0.0

# Sécurité
PyJWT>=2.8.0
cryptography>=41.0.0
